# perf(rust, python): speed up write_csv for time-zone-aware columns