--- a/polars/polars-time/src/windows/test.rs
+++ b/polars/polars-time/src/windows/test.rs
@@ -636,6 +636,7 @@ fn test_rolling_lookback() {
         None,
     )
     .unwrap(); // unwrapping as we pass None as the time zone
+    println!("dates: {:?}", dates);
 
     // full lookbehind
     let groups = groupby_values(